Plan driven versus agile. Final for now...
Don't really now if that's how it should be as I get stuck all the time.
When I start to write about some method I really want to explain it but
then again it is copying. So I tried to avoid all of this and ended up
with roughly 2 pages + 1 with references...